Assignment 9 - Complete
Tested user login and confirmed user are working, if you need to validate one of the users I created was JSmith with pw as password1

Confirmed attachments were loading and the download link was visible and working.

Checked that all data was showing in mysql. Users, tickets and attachments are all in the database.